Moved: Columns for NameWithFileIcon and FileCount to SharedColumns since Diff View also uses them
True story!
