+++ src/NexusMods.App.UI/Controls/TreeDataGrid/SharedColumns.cs
@@ -57,4 +57,47 @@ public static class SharedColumns
         public static string GetColumnHeader() => "Size";
         public static string GetColumnTemplateResourceKey() => ColumnTemplateResourceKey;
     }
+    
+    /// <summary>
+    /// Represents a file or folder name, accompanied by a file or folder icon.
+    /// </summary>
+    [UsedImplicitly]
+    public sealed class NameWithFileIcon : ICompositeColumnDefinition<NameWithFileIcon>
+    {
+        public static int Compare<TKey>(CompositeItemModel<TKey> a, CompositeItemModel<TKey> b) where TKey : notnull
+        {
+            var aValue = a.GetOptional<StringComponent>(StringComponentKey);
+            var bValue = b.GetOptional<StringComponent>(StringComponentKey);
+            return aValue.Compare(bValue);
+        }
+
+        public const string ColumnTemplateResourceKey = Prefix + "Name";
+        public static readonly ComponentKey StringComponentKey = ComponentKey.From(ColumnTemplateResourceKey + "_" + nameof(StringComponent));
+        public static readonly ComponentKey IconComponentKey = ComponentKey.From(ColumnTemplateResourceKey + "_" + nameof(UnifiedIconComponent));
+
+        public static string GetColumnHeader() => "Name";
+        public static string GetColumnTemplateResourceKey() => ColumnTemplateResourceKey;
+    }
+    
+    /// <summary>
+    /// Represents a count of files that lives under a folder. 
+    /// </summary>
+    [UsedImplicitly]
+    public sealed class FileCount : ICompositeColumnDefinition<FileCount>
+    {
+        public static int Compare<TKey>(CompositeItemModel<TKey> a, CompositeItemModel<TKey> b) where TKey : notnull
+        {
+            var aValue = a.GetOptional<ValueComponent<uint>>(ComponentKey);
+            var bValue = b.GetOptional<ValueComponent<uint>>(ComponentKey);
+            // Assuming ValueComponent has a comparable Value property or implements IComparable
+            // Adjust the comparison logic if ValueComponent comparison needs specific handling
+            return aValue.Compare(bValue); 
+        }
+
+        public const string ColumnTemplateResourceKey = Prefix + nameof(FileCount);
+        public static readonly ComponentKey ComponentKey = ComponentKey.From(ColumnTemplateResourceKey + "_" + "FileCount");
+
+        public static string GetColumnHeader() => "File Count";
+        public static string GetColumnTemplateResourceKey() => ColumnTemplateResourceKey;
+    }
 }
